Cerret Class - Hotboxes Added
Added 3 circular hotboxes to represent collision for the ferret class. There will be a hitbox that track the back, middle, and front of the ferret as ferret's are long objects and there is no rotated rectangle collision to be used for tracking. The plan is to use the front and back collisions to check for ground being walked on and all three hitboxes to check collision against thing slike walls.