Ferret Class Cleaned Up
Cleaned up the ferret class to now do away with (most) magic numbers in favor of magic numbers or allowing and now having the constructor dictate the location of the object. Also updated the documentation for the ferret class to reflect the current implementation.

Additionally, fixed a defect with the linker that was pointing to a wrong full fill path instead of the correct shortened local file path without using the entire filepath.